Do not jump on refresh
JSXTransformer was working by waiting for the dom to be loaded and then transform + eval all the script tags. This was bad for the experience because all the text would show up and the all the examples. If you were working on an example deep down, you'd lose your position all the time.

I fixed it by using babel-standalone and running the parsing + eval code right after the block was defined in a <script> tag that already existed. Also it's no longer using a super old transform and I moved to react-dom to remove a warning.